media/json: read JSON through the EBNF grammar, one policy per codec
`fjs/media/json/parser` is now the reader that PR #1908 built as
`fjs/ebnf/lib/json/parser`, moved here and generalized over the codec's
numeric policy: `mappings(policy)` is the rewrite set over the grammar in
`fjs/ebnf/lib/json`, its `number` mapping handing the policy the exact
lexeme, and `parse(policy)(text)` folds it into the LL(1) parse of
`[json, eof]`. Both codecs' `parse` run it, so the standard codec still
reads every number as JavaScript does and the extended codec still reads
bare integers as `bigint` and refuses `1e400` — the numeric-policy contract
`self-contained-tokenizer` held a value mapping could not keep. The
container state machine over a token stream is deleted, and the tokenizer
stays public and unchanged, off `parse`'s path.

A number the policy refuses is the policy's own message standing in the
number's place, a container holding one is that error too, and the first
in document order is the parse's; a parse that fails reports where —
`unexpected end`, or `unexpected symbol at N` — rather than which token.
The error type stays `string`, so every consumer of `json.parse` and
`extended.parse` is unchanged. `fjs/ebnf` no longer depends on
`fjs/media/json`.

`parser-container-stack-bookkeeping` described the deleted machine and is
closed; `self-contained-tokenizer` records that the value route landed as
the reader and what is left of it, the public `tokenize`.

Changelog:
- **BREAKING CHANGES:** `media/json/parser`: `parse(policy)` takes the text
  rather than a token list, `NumberPolicy<P>` is `(lexeme: string) =>
  Result<P, string>` rather than a function of a `NumberToken`, and the
  parser's state types are gone; the syntax error message names the index
  the parse failed at
- `media/json`: `parse` in both codecs reads through the JSON grammar of
  `ebnf/lib/json`, folded by `ebnf/ll1`; `ebnf/lib/json/parser` moved to
  `media/json/parser`
