+++ fjs/media/json/parser/types.ts
@@ -1,31 +1,26 @@
 /**
- * Types for the shared structural JSON parser: its numeric policy, the tree
- * that policy produces, and the parser's internal state.
+ * Type-level API of the JSON reader: the numeric policy a codec supplies,
+ * the tree that policy parses into, and the output alphabet the reader's
+ * mappings return into — the input alphabet is `../../../ebnf/utf16/types.ts`.
  *
  * @module
  */
 
 import type { Tree } from '../types.ts'
-import type { OrderedMap } from '../../../types/ordered_map/types.ts'
-import type { List } from '../../../types/list/types.ts'
 import type { Result } from '../../../types/result/types.ts'
-import type { JsonToken } from '../tokenizer/types.ts'
-import type { NumberToken } from '../../../js/tokenizer/types.ts'
-
-/** JSON tokens that carry a directly-usable value. */
-export type _ValueToken = Extract<JsonToken, { readonly kind: 'null' | 'false' | 'true' | 'string' | 'number' }>
 
 /**
- * A numeric policy: how one codec materializes a JSON number token into its
- * own numeric domain `P`.
+ * A numeric policy: how one codec materializes a JSON number into its own
+ * numeric domain `P`.
  *
- * The policy is handed the token itself, so it decides from the exact lexeme —
- * before any narrowing the parser might otherwise have imposed. It may fail:
- * a valid JSON number can be outside the domain the codec offers (the extended
- * codec has no non-finite `number`), and that has to be an ordinary parse
- * error rather than an escaping runtime exception.
+ * The policy is handed the number's exact lexeme, as the grammar admits it —
+ * `-?(0|[1-9][0-9]*)(\.[0-9]+)?([eE][+-]?[0-9]+)?` — before any narrowing the
+ * reader might otherwise have imposed. It may fail: a valid JSON number can
+ * be outside the domain the codec offers (the extended codec has no
+ * non-finite `number`), and that has to be an ordinary parse error, its
+ * message the policy's, rather than an escaping runtime exception.
  */
-export type NumberPolicy<P> = (token: NumberToken) => Result<P, string>
+export type NumberPolicy<P> = (lexeme: string) => Result<P, string>
 
 /**
  * The tree a numeric policy `P` parses into: JSON's containers over `P` plus
@@ -33,40 +28,17 @@ export type NumberPolicy<P> = (token: NumberToken) => Result<P, string>
  */
 export type ParseUnknown<P> = Tree<P | null | boolean | string>
 
-export type _JsonObject<P> = {
-    readonly kind: 'object'
-    readonly values: OrderedMap<ParseUnknown<P>>
-    readonly key: string
-}
+/**
+ * What the `string` rule's mapping returns: the string the rule spells,
+ * every escape decoded.
+ */
+export type Text = { readonly id: 'text', readonly value: string }
 
-export type _JsonArray<P> = {
-    readonly kind: 'array'
-    readonly values: List<ParseUnknown<P>>
-}
+/**
+ * What a value's mapping returns: the JSON value, or, where the policy
+ * refused a number under it, the policy's error in the value's place.
+ */
+export type Json<P> = { readonly id: 'json', readonly result: Result<ParseUnknown<P>, string> }
 
-type _JsonStackElement<P> = |
-    _JsonObject<P> |
-    _JsonArray<P>
-
-export type _JsonStack<P> = List<_JsonStackElement<P>>
-
-export type _StateParse<P> = {
-    readonly status: '' | '[' | '[v' | '[,' | '{' | '{k' | '{:' | '{v' | '{,'
-    readonly top: _JsonStackElement<P> | null
-    readonly stack: _JsonStack<P>
-}
-
-type _StateResult<P> = {
-    readonly status: 'result'
-    readonly value: ParseUnknown<P>
-}
-
-type _StateError = {
-    readonly status: 'error'
-    readonly message: string
-}
-
-export type _JsonState<P> = |
-    _StateParse<P> |
-    _StateResult<P> |
-    _StateError
+/** The output alphabet: what the mappings return. */
+export type Out<P> = Text | Json<P>
